+++ src/QueryMorphTo.php
@@ -22,9 +22,8 @@ namespace LaravelJsonApi\Eloquent;
 use Illuminate\Database\Eloquent\Model;
 use LaravelJsonApi\Contracts\Schema\Filter;
 use LaravelJsonApi\Contracts\Store\QueryOneBuilder;
-use LaravelJsonApi\Core\Query\QueryParameters;
+use LaravelJsonApi\Core\Query\Custom\ExtendedQueryParameters;
 use LaravelJsonApi\Eloquent\Fields\Relations\MorphTo;
-use LaravelJsonApi\Eloquent\Polymorphism\MorphParameters;
 use function is_null;
 
 class QueryMorphTo implements QueryOneBuilder
@@ -52,7 +51,7 @@ class QueryMorphTo implements QueryOneBuilder
     {
         $this->model = $model;
         $this->relation = $relation;
-        $this->queryParameters = new QueryParameters();
+        $this->queryParameters = new ExtendedQueryParameters();
     }
 
     /**
@@ -123,14 +122,13 @@ class QueryMorphTo implements QueryOneBuilder
     private function prepareResult(?Model $related): ?Model
     {
         if ($related) {
-            $parameters = new MorphParameters(
-                $schema = $this->relation->schemaFor($related),
-                $this->queryParameters,
-            );
+            $schema = $this->relation->schemaFor($related);
+            $parameters = $this->queryParameters->forSchema($schema);
 
             $schema
                 ->loaderFor($related)
-                ->loadMissing($parameters->includePaths());
+                ->loadMissing($parameters->includePaths())
+                ->loadCount($parameters->countable());
         }
 
         return $related;