+++ src/QueryToMany.php
@@ -29,7 +29,7 @@ use Illuminate\Support\Collection;
 use Illuminate\Support\LazyCollection;
 use LaravelJsonApi\Contracts\Pagination\Page;
 use LaravelJsonApi\Contracts\Store\QueryManyBuilder;
-use LaravelJsonApi\Core\Query\QueryParameters;
+use LaravelJsonApi\Core\Query\Custom\ExtendedQueryParameters;
 use LaravelJsonApi\Eloquent\Fields\Relations\ToMany;
 use LogicException;
 use function get_class;
@@ -39,6 +39,11 @@ class QueryToMany implements QueryManyBuilder
 {
 
     use HasQueryParameters;
+
+    /**
+     * @var Schema
+     */
+    private Schema $schema;
 
     /**
      * @var Model
@@ -53,14 +58,16 @@ class QueryToMany implements QueryManyBuilder
     /**
      * QueryToMany constructor.
      *
+     * @param Schema $schema
      * @param Model $model
      * @param ToMany $relation
      */
-    public function __construct(Model $model, ToMany $relation)
+    public function __construct(Schema $schema, Model $model, ToMany $relation)
     {
+        $this->schema = $schema;
         $this->model = $model;
         $this->relation = $relation;
-        $this->queryParameters = new QueryParameters();
+        $this->queryParameters = new ExtendedQueryParameters();
     }
 
     /**
@@ -146,6 +153,8 @@ class QueryToMany implements QueryManyBuilder
      */
     public function query(): JsonApiBuilder
     {
+        $this->prepareModel();
+
         $base = $this->relation->schema()->relatableQuery(
             $this->request, $this->getRelation()
         );
@@ -185,6 +194,20 @@ class QueryToMany implements QueryManyBuilder
             $name,
             get_class($this->model)
         ));
+    }
+
+    /**
+     * @return $this
+     */
+    private function prepareModel(): self
+    {
+        if ($this->relation->isCountableInRelationship()) {
+            $this->model->loadCount(
+                $this->relation->withCountName(),
+            );
+        }
+
+        return $this;
     }
 
 }