+++ tests/lib/Acceptance/Relations/MorphToMany/VideoTags/QueryTest.php
@@ -41,6 +41,9 @@ class QueryTest extends TestCase
             ->cursor();
 
         $this->assertTags($video->tags()->get(), $actual);
+
+        // as the relationship is countable, we expect the count to be loaded so the relationship meta is complete.
+        $this->assertEquals(count($actual), $video->tags_count);
     }
 
     public function testWithIncludePaths(): void